Added Travis-CI integration
Discovered that GNU Make uses timestamps to tell which files need
to be updated, which causes it to think it needs to rerun aclocal
or automake. Workaround for this is to give files the same timestamp:

    $ touch -rbuild/dummy config* aclocal* Makefile*

This generally seems to solve the mixup where a build fails from
time to time.